automated cache flag switch
Cache flag was always set to overwrite. Meaning: if there are more then 1 PBF files within the import directory the latter one will overwrite the first one, a bit of a waste of resoures.. Now the container autodetects if cache already exists and if so it appends. If you want to reimport and delete all data, just delete the cache contents.